grab_stock/pyproject.toml
[tool.poetry]
name = "grab-stock"
version = "0.1.0"
description = "20 NASDAQ stock grabber"
authors = ["Wesley So <wesleyso0614@gmail.com>"]
license = "PROPRIETARY"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.5.3"
requests = "^2.28.2"
python-dotenv = "^0.21.1"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
